consolidate api tests, use __mocks__ request
